requirements.txt
torch>=2.1
torchvision>=0.16
opencv-python>=4.8
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
pyyaml>=6.0
tqdm>=4.66
ultralytics>=8.0.200
transformers>=4.41
sentencepiece>=0.2.0
